Rename creator files
