--- src/components/creator/CreatorEditor.tsx
@@ -1,9 +1,0 @@
-import { Header } from "../header/Header"
-import Placeholder from "../../assets/placeholder.png"
-
-export const CreatorEditor = () =>{
-    return <>
-    <Header/>
-     <img alt={"neco"} src={`.${Placeholder}`}/>
-</>
-}
