+++ src/components/creator/Editor.tsx
@@ -1,0 +1,11 @@
+import { Header } from "../header/Header";
+import Placeholder from "../../assets/placeholder.png";
+
+export const CreatorEditor = () => {
+  return (
+    <>
+      <Header />
+      <img alt={"neco"} src={`.${Placeholder}`} />
+    </>
+  )
+}
